requirements.txt
Django
djangorestframework
django_extensions
django_annoying
django-model-utils
celery[redis]
django-celery-results
# django-celery-beat
path.py
iPython
